% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization.R
\name{addRowToTau}
\alias{addRowToTau}
\title{split group q of provided tau randomly into two into}
\usage{
addRowToTau(tau, q)
}
\arguments{
\item{tau}{provided tau}

\item{q}{indice of group to split}
}
\value{
new tau
}
\description{
split group q of provided tau randomly into two into
}
