% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{convertGroupPair}
\alias{convertGroupPair}
\title{transform a pair of block identifiers (q,l) into an identifying integer}
\usage{
convertGroupPair(q, l, Q, directed)
}
\arguments{
\item{q}{indicator of a latent block}

\item{l}{indicator of a latent block}

\item{Q}{number of latent blocks}

\item{directed}{indicates if the graph is directed}
}
\description{
this is the inverse function of convertGroupPairIdentifier()
}
