% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{convertNodePair}
\alias{convertNodePair}
\title{transform a pair of nodes (i,j) into an identifying integer}
\usage{
convertNodePair(i, j, n, directed)
}
\arguments{
\item{i}{scalar or vector}

\item{j}{scalar or vector, same length as i}

\item{n}{number of vertices}

\item{directed}{booelan to indicate whether the model is directed or undirected}
}
\description{
Associates an identifying integer with a pair of nodes (i,j)
}
\details{
returns the row number of the matrix build by listNodePairs(n)
    containing the pair (i,j)
}
