% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{emv_gamma}
\alias{emv_gamma}
\title{compute the MLE in the Gamma model using the Newton-Raphson method}
\usage{
emv_gamma(L, M, param.old, epsilon = 0.001, nb.iter.max = 10)
}
\arguments{
\item{L}{weighted mean of log(data)}

\item{M}{weighted mean of the data}

\item{param.old}{parameters of the Gamma distribution}

\item{epsilon}{threshold for the stopping criterion}

\item{nb.iter.max}{maximum number of iterations}
}
\value{
updated parameters of the Gamma distribution
}
\description{
compute the MLE in the Gamma model using the Newton-Raphson method
}
