% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{getBestQ}
\alias{getBestQ}
\title{optimal number of SBM blocks}
\usage{
getBestQ(bestSolutionAtQ)
}
\arguments{
\item{bestSolutionAtQ}{output of \code{fitNSBM()}, i.e. a list of estimation results for varying number of latent blocks}
}
\value{
a list the maximal value of the ICL criterion among the provided solutions along with the best number of latent blocks
}
\description{
returns the number of SBM blocks that maximizes the ICL
}
\examples{
# res_gauss is the output of a call of fitNSBM()
getBestQ(res_gauss)
}
