% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{getTauql}
\alias{getTauql}
\title{Evaluate tau_q*tau_l in the noisy stochastic block model}
\usage{
getTauql(q, l, tau, n, directed)
}
\arguments{
\item{q}{indicator of a latent block}

\item{l}{indicator of a latent block}

\item{tau}{variational parameters}

\item{n}{number of vertices}

\item{directed}{booelan to indicate whether the model is directed or undirected}
}
\description{
Evaluate tau_q*tau_l in the noisy stochastic block model
}
