% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization.R
\name{initialTau}
\alias{initialTau}
\title{compute intial values for tau}
\usage{
initialTau(Q, dataMatrix, nbOfTau, percentageOfPerturbation, directed)
}
\arguments{
\item{Q}{number of latent blocks in the noisy stochastic block model}

\item{dataMatrix}{observed dense adjacency matrix}

\item{nbOfTau}{number of initializations for the latent block memberships}

\item{percentageOfPerturbation}{percentage of node labels that are perturbed to obtain further initial points}

\item{directed}{booelan to indicate whether the model is directed or undirected}
}
\value{
a list of length nbOfTau of initial points for tau
}
\description{
returns a list of length nbOfTau of initial points for tau using spectral clustering with absolute values, kmeans and random perturbations of these points
}
