% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{J.gamma}
\alias{J.gamma}
\title{evaluate the objective in the Gamma model}
\usage{
J.gamma(param, L, M)
}
\arguments{
\item{param}{parameters of the Gamma distribution}

\item{L}{weighted mean of log(data)}

\item{M}{weighted mean of the data}
}
\value{
value of the lower bound of the log-likelihood function
}
\description{
evaluate the objective in the Gamma model
}
