% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{JEvalMstep}
\alias{JEvalMstep}
\title{evaluation of the objective in the Gauss model}
\usage{
JEvalMstep(VE, mstep, data, modelFamily, directed)
}
\arguments{
\item{VE}{list with variational parameters tau and rho}

\item{mstep}{list with current model parameters and additional auxiliary terms}

\item{data}{data vector in the undirected model, data matrix in the directed model}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}

\item{directed}{booelan to indicate whether the model is directed or undirected}
}
\value{
value of the ELBO and the complete log likelihood function
}
\description{
evaluation of the objective in the Gauss model
}
