% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{listNodePairs}
\alias{listNodePairs}
\title{returns a list of all possible node pairs (i,j)}
\usage{
listNodePairs(n, directed = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{directed}{indicates if the graph is directed}
}
\value{
a 2-column matrix with all possible node pairs (i,j)
}
\description{
returns a list of all possible node pairs (i,j)
}
