% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure.R
\name{lvaluesNSBM}
\alias{lvaluesNSBM}
\title{compute conditional l-values in the noisy stochastic block model}
\usage{
lvaluesNSBM(dataVec, Z, theta, directed = FALSE, modelFamily = "Gauss")
}
\arguments{
\item{dataVec}{data vector}

\item{Z}{a node clustering}

\item{theta}{list of parameters for a noisy stochastic block model}

\item{directed}{indicates if the graph is directed}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss} and \code{Gamma}}
}
\value{
conditional l-values in the noisy stochastic block model
}
\description{
compute conditional l-values in the noisy stochastic block model
}
