% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mat2Vect}
\alias{mat2Vect}
\title{mat2Vect : Builds a vector from a (symmetric) matrix}
\usage{
mat2Vect(M, symmetric, diag = FALSE)
}
\arguments{
\item{symmetric}{: TRUE is the network is not  directed. FALSE otherwise}

\item{diag}{: FALSE if the self interactions are neglected. TRUE otherwise (default value =  FALSE)}

\item{V}{: vector}
}
\value{
matrix fulfilled with the elements of the matrix
}
\description{
mat2Vect : Builds a vector from a (symmetric) matrix
}
\examples{
V <- c(0,15);
M <- vect2Mat(V, symmetric = TRUE, diag=FALSE)
V2 <- mat2Vect(M, symmetrice = TRUE, diag = FALSE)
}
