% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{n2nbPairs}
\alias{n2nbPairs}
\title{n2nbPairs : compute the number of Dyads from the number of nodes for a network}
\usage{
n2nbPairs(n, symmetric, diag = FALSE)
}
\arguments{
\item{n}{: number of nodes}

\item{symmetric}{: TRUE is the network is not  directed. FALSE otherwise}

\item{diag}{: FALSE if the self interactions are neglected. TRUE otherwise (default value =  FALSE)}
}
\value{
number of dyades
}
\description{
n2nbPairs : compute the number of Dyads from the number of nodes for a network
}
\examples{
n <- 10;
N <- n2nbPairs(n, symmetric = TRUE, diag=FALSE)
}
