% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{nbPairs2n}
\alias{nbPairs2n}
\title{nbPairs2n : compute the number of nodes of a network from its number of dyads}
\usage{
nbPairs2n(N, symmetric, diag = FALSE)
}
\arguments{
\item{N}{: number of dyads}

\item{symmetric}{: TRUE is the network is not  directed. FALSE otherwise}

\item{diag}{: FALSE if the self interactions are neglected. TRUE otherwise (default value =  FALSE)}
}
\value{
number of dyades
}
\description{
nbPairs2n : compute the number of nodes of a network from its number of dyads
}
\examples{
N <- 45;
n <- nbPairs2n(n, symmetric = TRUE, diag=FALSE)
}
