% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoisySBM-package.R
\docType{package}
\name{NoisySBM-package}
\alias{NoisySBM}
\alias{NoisySBM-package}
\title{NoisySBM: Inference of a network topology from a one or several noisy versions of the network}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stéphane Robin \email{stephane.robin@inrae.fr} (\href{https://orcid.org/0000-0003-1045-069X}{ORCID})

}
\keyword{internal}
