% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{plotICL}
\alias{plotICL}
\title{plot ICL curve}
\usage{
plotICL(res)
}
\arguments{
\item{res}{output of fitNSBM()}
}
\value{
figure of ICL curve
}
\description{
plot ICL curve
}
\examples{
# res_gauss is the output of a call of fitNSBM()
plotICL(res_gauss)
}
