% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure.R
\name{q_delta_ql}
\alias{q_delta_ql}
\title{auxiliary function for the computation of q-values}
\usage{
q_delta_ql(theta, ind, t, modelFamily = "Gauss")
}
\arguments{
\item{theta}{list of parameters for a noisy stochastic block model}

\item{ind}{indicator for a pair of latent blocks}

\item{t}{l-values}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss} and \code{Gamma}}
}
\description{
auxiliary function for the computation of q-values
}
