% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure.R
\name{qvaluesNSBM}
\alias{qvaluesNSBM}
\title{compute q-values in the noisy stochastic block model}
\usage{
qvaluesNSBM(
  dataVec,
  Z,
  theta,
  lvalues,
  modelFamily = "Gauss",
  directed = FALSE
)
}
\arguments{
\item{dataVec}{data vector}

\item{Z}{a node clustering}

\item{theta}{list of parameters for a noisy stochastic block model}

\item{lvalues}{conditional l-values in the noisy stochastic block model}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss} and \code{Gamma}}

\item{directed}{indicates if the graph is directed}
}
\value{
q-values in the noisy stochastic block model
}
\description{
compute q-values in the noisy stochastic block model
}
