% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_exp}
\alias{res_exp}
\title{Output of fitNSBM() on a dataset applied in the exponential NSBM}
\format{
List with estimation results for different number of SBM blocks. Output of fitNSBM()
}
\usage{
res_exp
}
\description{
Parameter estimates fitted on a dataset given in the vignette
}
\keyword{datasets}
