% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_gauss}
\alias{res_gauss}
\title{Output of fitNSBM() on a dataset applied in the Gaussian NSBM}
\format{
List with estimation results for different number of SBM blocks. Output of fitNSBM()
}
\usage{
res_gauss
}
\description{
Parameter estimates fitted on a dataset given in the vignette
}
\keyword{datasets}
