% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNoisySBM.R
\name{rNoisySBM}
\alias{rNoisySBM}
\title{Simulate Noisy SBM}
\usage{
rNoisySBM(
  nbNodes,
  directed,
  blockProp,
  connectParam,
  emissionParam,
  seed = NULL
)
}
\arguments{
\item{nbNodes}{: number of nodes}

\item{directed}{: directed or not (directed = TRUE or FALSE)}

\item{blockProp}{: proportions of the Blocks. Vector of size nbBlocks}

\item{connectParam}{: probabilities of connections inside and between blocks. Matrix of size nbBlocks.}

\item{emissionParam}{: parameters of the emission of the Noisy SBM. List of two terms : noEdgeParam and edgeParam. Each element is a list of means (vector or size nbScores) and variance matrix (square matrix of size nbScores)}

\item{seed}{: set the seed for the random simulation (default value  = NULL)}
}
\value{
A list of containing the noisy networks (noisyNetworks a list of length nbScores of matrices of dimension nbNodes x nbNodes), the true underlying network (trueNetwork) and the clustering of the nodes (memberships)
}
\description{
\code{rNoisyMBM} simulates a collection of networks which are noisy versions of un underlying network.  This network may be directed or not. See vignette for more informations
}
\examples{
nbNodes  <- 100
directed <- TRUE
blockProp <- c(1/3,1/2,1/6)
nbBlocks   <- length(blockProp)
connectParam <- matrix(rbeta(nbBlocks^2,1.5,1.5 ),nbBlocks,nbBlocks)
connectParam <- 0.5*(connectParam + t(connectParam))
emissionParam <- list()
nbScores <- 4
emissionParam$noEdgeParam <- list(mean=rep(0,nbScores));
emissionParam$noEdgeParam$var <- diag(0.1,nrow = nbScores,ncol = nbScores)
emissionParam$EdgeParam <- list( mean= 1:nbScores)
emissionParam$EdgeParam$var <-  diag(0.1,nrow = nbScores,ncol = nbScores)
dataSim <- rNoisySBM(nbNodes,directed = TRUE, blockProp,connectParam,emissionParam,seed = NULL)

}
