% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization.R
\name{spectralClustering}
\alias{spectralClustering}
\title{spectral clustering with absolute values}
\usage{
spectralClustering(A, K)
}
\arguments{
\item{A}{adjacency matrix}

\item{K}{number of desired clusters}
}
\value{
a vector containing a node clustering into K groups
}
\description{
performs absolute spectral clustering of an adjacency matrix
}
