% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization.R
\name{tauDown}
\alias{tauDown}
\title{Create new initial values by merging pairs of groups of provided tau}
\usage{
tauDown(tau, nbOfMerges)
}
\arguments{
\item{tau}{soft node clustering}

\item{nbOfMerges}{number of required merges of blocks}
}
\value{
a list of length nbOfMerges (at most) of initial points for tau
}
\description{
Create nbOfMerges new initial values by merging nbOfMerges (or all possible) pairs of groups of provided tau
}
