% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization.R
\name{tauUp}
\alias{tauUp}
\title{Create new values of tau by splitting groups of provided tau}
\usage{
tauUp(tau, nbOfSplits = 1)
}
\arguments{
\item{tau}{soft node clustering}

\item{nbOfSplits}{number of required splits of blocks}
}
\value{
a list of length nbOfSplits (at most) of initial points for tau
}
\description{
Create nbOfSplits (or all) new values of tau by splitting nbOfSplits (or all) groups of provided tau
}
