% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{tauUpdate}
\alias{tauUpdate}
\title{Compute one iteration to solve the fixed point equation in the VE-step}
\usage{
tauUpdate(tau, log.w, log.1mw, data, VE, mstep, modelFamily, directed)
}
\arguments{
\item{tau}{current value of tau}

\item{log.w}{value of log(w)}

\item{log.1mw}{value of log(1-w)}

\item{data}{data vector in the undirected model, data matrix in the directed model}

\item{VE}{list with variational parameters tau and rho}

\item{mstep}{list with current model parameters and additional auxiliary terms}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}

\item{directed}{booelan to indicate whether the model is directed or undirected}
}
\value{
updated value of tau
}
\description{
Compute one iteration to solve the fixed point equation in the VE-step
}
