% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm.R
\name{VEstep}
\alias{VEstep}
\title{VE-step}
\usage{
VEstep(VE, mstep, data, modelFamily, directed, fix.iter = 5)
}
\arguments{
\item{VE}{list with variational parameters tau and rho}

\item{mstep}{list with current model parameters and additional auxiliary terms}

\item{data}{data vector in the undirected model, data matrix in the directed model}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}

\item{directed}{booelan to indicate whether the model is directed or undirected}

\item{fix.iter}{maximal number of iterations for fixed point equation}
}
\value{
updated list \code{VE} with variational parameters tau and rho
}
\description{
performs one VE-step, that is, update of tau and rho
}
