% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNoisySBM.R
\name{estimateNoisySBM}
\alias{estimateNoisySBM}
\title{Inference of the Noisy SBM  Model}
\usage{
estimateNoisySBM(
  scoreList,
  directed = FALSE,
  estimOptions = list(),
  monitoring = list()
)
}
\arguments{
\item{directed}{:  if true the inference network is directed. Default value = FALSE.}

\item{data}{:  list of d Noisy versions of a network}
}
\value{
a list of estimated parameters for the different models ordered by decreasing ICL. If save \code{= FALSE}, the length is of length 1 (only the better model is returned).
\describe{
}
}
}
}
\description{
Inference of the Noisy SBM  Model
}
\details{
The function \code{multipartiteBM} selects the better numbers of blocks in each FG (with a penalized likelihood criterion). The model selection is performed with a forward backward strategy and the likelihood of each model is maximized with a variational EM).
}
\examples{
nNodes <- 100
}
