% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNoisySBM.R
\name{rNoisySBM}
\alias{rNoisySBM}
\title{Simulate Noisy SBM}
\usage{
rNoisySBM(
  nNodes,
  symmetric,
  mixtureParam,
  connectParam,
  emissionParam,
  seed = NULL
)
}
\arguments{
\item{nNodes}{: number of nodes}

\item{symmetric}{: symmetric (directed or not) (symmetric = TRUE or FALSE)}

\item{mixtureParam}{: proportions of the Blocks. Vector of size K}

\item{connectParam}{: probabilities of connections inside and between blocks. Matrix of size K.}

\item{emissionParam}{: parameters of the emission of the Noisy SBM. List of two terms : noEdgeParam and edgeParam. Each element is a list of means (vector or size d) and variance matrix (square matrix of size d)}

\item{seed}{: set the seed for the random simulation (default value  = NULL)}
}
\value{
A list of lists containing the noisy networks (noisyNetworks), the true underlying network trueNetwork and
Each element of  list_net corresponds to a network : each network is a list containing  the matrix (mat) , the type of network(diradj, adj, inc), the functional group in row (rowFG) and the functional group in columns (colFG)
}
\description{
\code{rNoisyMBM} simulates a collection of networks which are noisy versions of un underlying network.  This network may be directed or not. See vignette for more informations
}
\examples{
nNodes  = 100
directed = TRUE
mixtureParam = c(1/3,1/2,1/6)
nBlocks   = length(mixtureParam)
connectParam <- matrix(rbeta(nBlocks^2,1.5,1.5 ),nBlocks,nBlocks)
connectParam <- 0.5*(connectParam + t(connectParam))
emissionParam <- list()
d <- 4
emissionParam$noEdgeParam = list(mean=rep(0,d),var = diag(0.1,nrow = d,ncol = d))
emissionParam$EdgeParam = list( mean= 1:d,var =  diag(0.1,nrow = d,ncol = d))
dataSim <- rNoisySBM(nNodes,symmetric = TRUE, mixtureParam,connectParam,emissionParam,seed = NULL)

}
